zadanie domowe - poprawione i sprawdzone

+++ src/main/java/pl/sda/bikerepositoryapp/controller/BikeController.java
@@ -4,6 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import pl.sda.bikerepositoryapp.model.Bike;
@@ -29,12 +30,18 @@ public class BikeController {
     }
 
     @GetMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public @ResponseBody
-    Bike getBike(@PathVariable String id) {
-        LOG.info("Controller: Displaying specific bike.");
+    public ResponseEntity<Bike> getBike(@PathVariable String id) {
         int index = Integer.parseInt(id);
-        return bikeService.findBike(index);
+        if (index == 418) {
+            return ResponseEntity.status(HttpStatus.I_AM_A_TEAPOT).build();
+        }
+        Bike bike = bikeService.findBike(index);
+        if (bike == null) {
+            LOG.error("Controller: No bike with index " + index + " found.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        LOG.info("Controller: Displaying specific bike.");
+        return ResponseEntity.ok(bike);
     }
 
     @PostMapping
@@ -45,9 +52,29 @@ public class BikeController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void removeBike(@PathVariable String id) {
-        LOG.info("Controller: Removing specific bike.");
+    public ResponseEntity<Bike> removeBike(@PathVariable String id) {
         int index = Integer.parseInt(id);
+        if (index == 418) {
+            return ResponseEntity.status(HttpStatus.I_AM_A_TEAPOT).build();
+        }
+        if (bikeService.findBike(index) == null) {
+            LOG.error("Controller: No bike with index " + index + " have been found.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        LOG.info("Controller: Removing specified bike.");
         bikeService.deleteBike(index);
+        return null;
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public void handleIllegalArgument() {
+        LOG.error("Resolving IllegalArgumentException.");
+    }
+
+    @ExceptionHandler(NumberFormatException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public void handleFormatException() {
+        LOG.error("Resolving NumberFormatException.");
     }
 }
