+++ src/main/java/pl/sda/bikerepositoryapp/service/BikeService.java
@@ -24,7 +24,12 @@ public class BikeService {
 
     public Bike findBike(int index) {
         LOG.info("Service: Searching for a bike by it's Index.");
-        return bikeRepository.findBike(index);
+        return bikeRepository.findBike(index).orElseThrow(() -> {
+                    String message = "No bike with id of \"" + index + "\" have been found.";
+                    LOG.error(message);
+                    return new IllegalArgumentException(message);
+                }
+        );
     }
 
     public void addBike(Bike bike) {
@@ -33,7 +38,12 @@ public class BikeService {
     }
 
     public void deleteBike(int index) {
-        LOG.info("Service: Deleting bike from listing.");
+        LOG.warn("Service: Deleting bike from listing.");
+        if(bikeRepository.deleteBike(index) == null) {
+            String message = "No dbke with id of \"" + index + "\" have been found.";
+            LOG.error(message);
+            throw new IllegalArgumentException(message);
+        }
         bikeRepository.deleteBike(index);
     }
 }
